Clean up tests and extract common DB mocking functionality to parent class
